KC Updated

+++ pwa-loc/scripts/services/keycloakManager.js
@@ -2,6 +2,7 @@ function KeycloakManager() {};
 
 
 KeycloakManager.OFFLINE_TIMEOUT = 3 * 60; // 3 mins, Overwritten by Config 'offlineTimeoutSec'
+KeycloakManager.RENEW_ACCESS_TOKEN_BEFORE_EXPIRED_TIME = 5; // seconds
 
 KeycloakManager.btnKeyCloakLogOutTag;
 KeycloakManager.keycloakMsgTag;
@@ -10,7 +11,6 @@ KeycloakManager.dialogTag;
 KeycloakManager.keycloakObj;
 KeycloakManager.timeSkew = 1;
 KeycloakManager.accessTokenTimeoutObj;
-KeycloakManager.refreshTokenTimeoutObj;
 KeycloakManager.offlineExpiredIntervalObj;
 
 // Flag set when user disabled or session expired.  OfflineTimeExpire checks this to not show msg when this is true.
@@ -41,8 +41,15 @@ KeycloakManager.getStatusSummary = function()
 {
 	var statusJson = {};
 
+	var processingAction = KeycloakLSManager.getProcessingAction();
+	statusJson.kcObjCreated = ( KeycloakManager.keycloakObj != undefined );
+	statusJson.isTokensExisted = ( KeycloakLSManager.getAccessToken() && processingAction != KeycloakLSManager.KEY_PROCESSING_ACTION_LOGOUT ) ? true: false;
+
+
+
 	statusJson.isAppOnline = ConnManagerNew.isAppMode_Online();
 
+	// TODO: Need to think about 'isKeycloakAuth'
 	statusJson.isKeycloakAuth = ( KeycloakLSManager.getAccessToken() ) ? true: false;
 	
 	statusJson.processingAction = KeycloakLSManager.getProcessingAction();
@@ -52,6 +59,20 @@ KeycloakManager.getStatusSummary = function()
 
 	statusJson.isLoggedIn = SessionManager.getLoginStatus();
 
+	// keycloak information
+	statusJson.kc = {
+		accessToken: KeycloakLSManager.getAccessToken(),
+		accessTokenParsed: KeycloakLSManager.getAccessTokenParsed(),
+		accessTokenValidInSeconds: KeycloakManager.calculateTokenExpiredTimeRemains(KeycloakLSManager.getAccessTokenParsed()),
+
+		refreshToken: KeycloakLSManager.getRefreshToken(),
+		refreshTokenParsed: KeycloakLSManager.getRefreshTokenParsed(),
+		refreshTokenValidInSeconds: KeycloakManager.calculateTokenExpiredTimeRemains(KeycloakLSManager.getRefreshTokenParsed()),
+
+		idToken: KeycloakLSManager.getIdToken(),
+		idTokenParsed: KeycloakLSManager.getIdTokenParsed()
+
+	};
 	return statusJson;
 };
 
@@ -83,27 +104,70 @@ KeycloakManager.setUpOnlineMode = function()
 	});
 
 	var statusJson = KeycloakManager.getStatusSummary();
-	// 	- If Not Authenticated, We can Authenticate..
-	if ( !statusJson.isKeycloakAuth ) KeycloakManager.authenticate();
-	else if ( statusJson.isKeycloakAuth )
+	if( !statusJson.kcObjCreated )
 	{
-		// NOTE: On Manual KeyCloak 'logout' click, we use this - saving 'process_action_logout' in localStorage before refrehs Page.
-		//		- We do not delete the keycloak auth data on 'logout' click - because we want to delete after confirm logout in server.
-		//		- When we enter website with this 'action_logout', 
-		//			- We need to remove 'processingAction' in localStorage
-		//			- on Authenticate, 
-		if ( statusJson.processingAction == KeycloakLSManager.KEY_PROCESSING_ACTION_LOGOUT ) KeycloakManager.authenticate();
-		else
+		if ( statusJson.isTokensExisted ) 
 		{
-			KeycloakManager.renewAccessToken(function(auth, errMsg) 
+			KeycloakManager.authenticate_WithToken(function(auth, errMsg) 
 			{
-				if ( auth ) KeycloakManager.authenticateSuccessActions();
-				else KeycloakManager.logout(); // CASES: Disabled User, Session/RefreshToken Expired
+				if ( auth ) {
+					KeycloakManager.eventMsg('Authenticated and Updated the access token.');
+					KeycloakManager.authenticateSuccessActions();
+				}
+				else 
+				{
+					KeycloakManager.logout(); // CASES: Disabled User, Session/RefreshToken Expired
+				}
 			});
-
-		}
+		} 
+		else 
+		{
+			KeycloakManager.authenticate();
+		} 
 	}
+	else
+	{
+		KeycloakManager.restartServiceToUpdateAccessToken();
+	}
+
 }
+
+// KeycloakManager.setUpOnlineMode = function()
+// {
+// 	// Enable the "Keyclock logout" button in the bottom & Set "logout" function for click button
+// 	KeycloakManager.btnKeyCloakLogOutTag.prop('disabled', false).html("AuthOut").show().off("click").click( () => { 
+// 		KeycloakManager.logout();
+// 	});
+
+// 	var statusJson = KeycloakManager.getStatusSummary();
+// 	// 	- If Not Authenticated, We can Authenticate..
+// 	if ( !statusJson.isKeycloakAuth ) KeycloakManager.authenticate();
+// 	else if ( statusJson.isKeycloakAuth )
+// 	{
+// 		// NOTE: On Manual KeyCloak 'logout' click, we use this - saving 'process_action_logout' in localStorage before refrehs Page.
+// 		//		- We do not delete the keycloak auth data on 'logout' click - because we want to delete after confirm logout in server.
+// 		//		- When we enter website with this 'action_logout', 
+// 		//			- We need to remove 'processingAction' in localStorage
+// 		//			- on Authenticate, 
+// 		if ( statusJson.processingAction == KeycloakLSManager.KEY_PROCESSING_ACTION_LOGOUT ) KeycloakManager.authenticate();
+// 		else if( KeycloakManager.keycloakObj == null )
+// 		{
+
+// 			KeycloakManager.authenticate_WithToken(function(auth, errMsg) 
+// 			{
+// 				if ( auth ) {
+// 					KeycloakManager.eventMsg('Authenticated and create a new access token.');
+// 					KeycloakManager.authenticateSuccessActions();
+// 				}
+// 				else KeycloakManager.logout(); // CASES: Disabled User, Session/RefreshToken Expired
+// 			});
+// 		}
+// 		else
+// 		{
+// 			KeycloakManager.restartServiceToUpdateAccessToken();
+// 		}
+// 	}
+// }
 
 // ---------------------------------------------------------------------------------------------------------
 // For OFFLINE appMode setup
@@ -133,13 +197,11 @@ KeycloakManager.authenticate = function()
 {
 	if( KeycloakManager.keycloakObj == undefined )
 	{
-		KeycloakLSManager.setLastKeycloakEvent("KeycloakManager.authenticate --- STARTING");
 		KeycloakManager.keycloakObj = KeycloakManager.createKeycloakObj();
 		KeycloakManager.setUpKeycloakObjEvents(KeycloakManager.keycloakObj);
 
 		// Authenticate
 		KeycloakManager.keycloakObj.init({ 
-			// onLoad: 'check-sso', // "check-sso", "login-required"
 			checkLoginIframe: false, 
 			scope: 'openid offline_access',
 			adapter: 'default',
@@ -148,11 +210,9 @@ KeycloakManager.authenticate = function()
 		{
 			if( !authenticated ) {
 				KeycloakManager.authenticateFailureActions();
-				KeycloakLSManager.setLastKeycloakEvent("KeycloakManager.authenticate --- False - Login now");
 				KeycloakManager.keycloakObj.login();
 			} 
 			else {
-				KeycloakLSManager.setLastKeycloakEvent("KeycloakManager.authenticate --- DONE");
 				KeycloakManager.eventMsg('Authenticated.');
 
 				MsgManager.msgAreaShowOpt( "Login with Keycloak success !", { cssClasses: 'notifDark', hideTimeMs: 2000 } );
@@ -160,11 +220,11 @@ KeycloakManager.authenticate = function()
 				// Save the username info..
 				var userName = KeycloakManager.keycloakObj.tokenParsed.preferred_username.toUpperCase();
 				var preLoginUser = AppInfoLSManager.getUserName();
-
-				if ( userName != preLoginUser ) 
+				
+				if ( preLoginUser != undefined && userName != preLoginUser ) 
 				{
-					KeycloakManager.eventMsg("User Changed, Deleting previous user data.");
-					
+					KeycloakManager.authenticateSuccessActions();
+
 					DataManager2.deleteAllStorageData(function () 
 					{
 						KeycloakManager.eventMsg( 'User Changed, Deleted previous user data.' ); 
@@ -173,16 +233,19 @@ KeycloakManager.authenticate = function()
 						AppInfoLSManager.setUserName(userName);
 						if ( SessionManager.cwsRenderObj ) SessionManager.cwsRenderObj.loadSavedUserName();
 
-						KeycloakManager.authenticateSuccessActions();
-
 						AppUtil.appReloadWtMsg("User Change - Deleteting previous user data ..");
 					});
 				}
 				else
 				{
 					KeycloakManager.authenticateSuccessActions();
+
+					// Save the new username - in case preLoginUser is undefined
+					if ( preLoginUser == undefined ){
+						AppInfoLSManager.setUserName(userName);
+						if ( SessionManager.cwsRenderObj ) SessionManager.cwsRenderObj.loadSavedUserName();
+					}
 				}
-				
 			}
 		})
 		.catch(function( errMsg ) {
@@ -217,17 +280,13 @@ KeycloakManager.createKeycloakObj = function()
 
 KeycloakManager.setUpKeycloakObjEvents = function( kcObj ) 
 {
-	// ----------------------
-	// Triggers below will RELOAD the browser
-
     kcObj.onAuthError = () => {
 		KeycloakManager.eventMsg("Auth Error");
 		KeycloakLSManager.setLastKeycloakEvent("onAuthError");
 
-
 		if( !KeycloakManager.errorTriggerCalled )
 		{
-			KeycloakManager.showDialog("User needs to authenticate.", kcObj.login );
+			KeycloakManager.showDialog("User needs to authenticate.", KeycloakManager.logout );
 			KeycloakManager.errorTriggerCalled = true;
 		}
 	}
@@ -239,25 +298,25 @@ KeycloakManager.setUpKeycloakObjEvents = function( kcObj )
 
 		if( !KeycloakManager.errorTriggerCalled )
 		{
-			KeycloakManager.showDialog("User needs to authenticate.", kcObj.login );
+			KeycloakManager.showDialog("User needs to authenticate.", KeycloakManager.logout );
 			KeycloakManager.errorTriggerCalled = true;
 		}
 	}
+
     kcObj.onAuthLogout = () => {
 		KeycloakManager.eventMsg('Auth Logout');
 		KeycloakLSManager.setLastKeycloakEvent("onAuthLogout");
 		
 		if( !KeycloakManager.errorTriggerCalled )
 		{
-			KeycloakManager.showDialog("User needs to authenticate.", kcObj.login );
+			KeycloakManager.showDialog("User needs to authenticate.", KeycloakManager.logout );
 			KeycloakManager.errorTriggerCalled = true;
 		}
 	}
 
 
-	// ----------------------
-	// Triggers below will NOT RELOAD the browser
 	// ----------------------------------------
+
 	kcObj.onAuthSuccess = () => {
 		KeycloakManager.eventMsg('Auth Success');  
 		KeycloakLSManager.setLastKeycloakEvent("onAuthSuccess");
@@ -271,7 +330,6 @@ KeycloakManager.setUpKeycloakObjEvents = function( kcObj )
    kcObj.onTokenExpired = () => {
 		KeycloakManager.eventMsg('Access token expired.');
 		KeycloakLSManager.setLastKeycloakEvent("onTokenExpired");
-		KeycloakManager.renewAccessToken();
 	}
 };
 
@@ -281,6 +339,9 @@ KeycloakManager.authenticateSuccessActions = function()
 	KeycloakLSManager.setLastKeycloakEvent("KeycloakManager.authenticateSuccessActions");
 	// Save tokens in Local Storage
 	KeycloakLSManager.setKeycloakInfo( KeycloakManager.keycloakObj );
+
+	// Turn on server to update Access Token
+	KeycloakManager.restartServiceToUpdateAccessToken();
 
 	// Turn on services to check Offline timeout
 	KeycloakManager.restartServiceToCheckOfflineTimeout();
@@ -301,41 +362,71 @@ KeycloakManager.authenticateFailureActions = function()
 // For renew Token
 
 // There is some case, the token expired methods called many times
-KeycloakManager.renewAccessToken = function(returnFunc)
+KeycloakManager.authenticate_WithToken = function(returnFunc)
 {
-	if( !KeycloakManager.processingNewAccessToken )
-	{
-		KeycloakLSManager.setLastKeycloakEvent("KeycloakManager.renewAccessToken --- STARTING");
-		KeycloakManager.processingNewAccessToken = true;
+	const accessToken = KeycloakLSManager.getAccessToken();
+	const refreshToken = KeycloakLSManager.getRefreshToken();
+	const idToken = KeycloakLSManager.getIdToken();
 
-		const accessToken = KeycloakLSManager.getAccessToken();
-		const refreshToken = KeycloakLSManager.getRefreshToken();
-		const idToken = KeycloakLSManager.getIdToken();
+	KeycloakManager.keycloakObj = KeycloakManager.createKeycloakObj();
+	KeycloakManager.setUpKeycloakObjEvents(KeycloakManager.keycloakObj);
 
-		KeycloakManager.keycloakObj = KeycloakManager.createKeycloakObj();
-		KeycloakManager.setUpKeycloakObjEvents(KeycloakManager.keycloakObj);
-
-		KeycloakManager.keycloakObj.init({
-			// onLoad: "check-sso", // "check-sso", "login-required"
-			checkLoginIframe: false,
-			token: accessToken,
-			refreshToken: refreshToken,
-			idToken: idToken,
-			timeSkew: KeycloakManager.timeSkew
-		}).then(function(authenticated) {
-			KeycloakManager.eventMsg("A new access token is created");
-			KeycloakLSManager.setLastKeycloakEvent("KeycloakManager.renewAccessToken --- DONE");
-			KeycloakManager.processingNewAccessToken = false;
-			if( returnFunc ) returnFunc(authenticated);
-		})
-		.catch(function( errMsg ) {
-			KeycloakManager.eventMsg("Failed to create a new access token.");
-			KeycloakLSManager.setLastKeycloakEvent("KeycloakManager.renewAccessToken --- FAILURE");
-			KeycloakManager.processingNewAccessToken = false;
-			if( returnFunc ) returnFunc(false, errMsg);
-		});
-	}
+	KeycloakManager.keycloakObj.init({
+		checkLoginIframe: false,
+		token: accessToken,
+		refreshToken: refreshToken,
+		idToken: idToken,
+		timeSkew: KeycloakManager.timeSkew
+	}).then(function(authenticated) {
+		if( returnFunc ) returnFunc(authenticated);
+	})
+	.catch(function( errMsg ) {
+		KeycloakManager.eventMsg("Failed to create a new access token.");
+		if( returnFunc ) returnFunc(false, errMsg);
+	});
 };
+
+KeycloakManager.restartServiceToUpdateAccessToken = function()
+{
+	// Stop the service to check Offline Timeout if it is started before
+	clearTimeout(KeycloakManager.accessTokenTimeoutObj);
+
+	// Start service again
+	var accessTokenParsed = KeycloakLSManager.getAccessTokenParsed();
+	var timeoutSeconds = KeycloakManager.calculateTokenExpiredTimeRemains(accessTokenParsed) - KeycloakManager.RENEW_ACCESS_TOKEN_BEFORE_EXPIRED_TIME;
+	KeycloakManager.accessTokenTimeoutObj = setTimeout(() => {
+		KeycloakManager.updateToken();
+	},timeoutSeconds * 1000);
+
+};
+
+KeycloakManager.updateToken = function()
+{
+	alert("UpdateToken");
+	KeycloakManager.keycloakObj.updateToken(-1).then(function(refreshed) {
+		KeycloakManager.eventMsg('Token is valid for ' + Math.round(KeycloakManager.keycloakObj.tokenParsed.exp + KeycloakManager.keycloakObj.timeSkew - new Date().getTime() / 1000) + ' seconds');
+		if (refreshed) {
+			KeycloakLSManager.setKeycloakInfo( KeycloakManager.keycloakObj );
+			KeycloakManager.restartServiceToUpdateAccessToken();
+			KeycloakManager.eventMsg("Access token is updated.");
+		}
+		else
+		{
+			KeycloakManager.eventMsg("Access token is not updated.");
+		}
+	}).catch(function(err) {
+		KeycloakManager.eventMsg('Failed to update access token');
+	});
+}
+
+KeycloakManager.calculateTokenExpiredTimeRemains = function(tokenParsed)
+{
+	if( tokenParsed.exp != undefined )
+	{
+		return Math.round(tokenParsed.exp + KeycloakManager.timeSkew - new Date().getTime() / 1000); // seconds
+	}
+	return;
+}
 
 // For Offline Timeout service
 KeycloakManager.restartServiceToCheckOfflineTimeout = function()
@@ -380,6 +471,7 @@ KeycloakManager.stopServiceToCheckOfflineTimeOut = function()
 	clearInterval(KeycloakManager.offlineExpiredIntervalObj);
 }
 
+
 // ==============================================================================
 // Logout
 
@@ -407,7 +499,6 @@ KeycloakManager.logout = function()
 		window.location.replace(logoutUrl);	
 	}
 };
-
 
 
 // ==============================================================================
@@ -440,6 +531,7 @@ KeycloakManager.calculateOfflineTimeRemains = function()
 	return KeycloakManager.OFFLINE_TIMEOUT - diffTimes;
 }
 
+
 // ==============================================================================
 // Other methods
 
@@ -462,6 +554,7 @@ KeycloakManager.getUserInfo = function()
 
 	return userInfo;
 };
+
 
 // ==============================================================================
 // Supportive methods
